+++ src/Mapping.sol
@@ -5,20 +5,20 @@ pragma solidity ^0.8.19;
 // and retrieve data efficiently. It is similar to a dictionary or hash map
 // in other programming languages.
 contract Mapping {
-    mapping(address => uint) public balances;
+    mapping(address => uint256) public balances;
 
-    function getBalance() external view returns (uint) {
+    function getBalance() external view returns (uint256) {
         return balances[msg.sender];
     }
 
-    function setBalance(uint _amount) external {
+    function setBalance(uint256 _amount) external {
         if (balances[msg.sender] != 0) {
             revert("Already set");
         }
         balances[msg.sender] = _amount;
     }
 
-    function updateBalance(uint _amount) external {
+    function updateBalance(uint256 _amount) external {
         balances[msg.sender] += _amount;
     }
